novelinsights/backend/pyproject.toml
[tool.poetry]
name = "novelinsights"
version = "0.1.0"
description = "AI-powered platform for tracking and summarizing stories"
authors = ["Jimmy Yao <jinghaoyao@gmail.com>"]
readme = "../../README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
llama-index = "^0.12.2"
llama-index-core = "^0.12.3"
llama-index-readers-file = "^0.4.1"
ebook-tree-parser = "^0.1.2"
unidecode = "^1.3.8"
python-dotenv = "^1.0.1"
llama-index-llms-openai = "^0.3.2"
llama-index-llms-anthropic = "^0.5.0"
lxml = "^5.3.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pgvector = "^0.3.6"
sentence-transformers = "^3.4.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
mypy = "^1.13.0"
alembic = "^1.14.0"
ipython = "^8.30.0"
ipykernel = "^6.29.5"
alembic-postgresql-enum = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
